smaccmpilot: Update to new Ivory Ix and loop functions.

+++ smaccmpilot/include/smaccmpilot/param.h
@@ -11,7 +11,7 @@ extern "C" {
 struct param_info {
     uint8_t param_type;
     char param_name[32U];
-    uint16_t param_index;
+    int32_t param_index;
     uint8_t* param_ptr_u8;
     uint16_t* param_ptr_u16;
     uint32_t* param_ptr_u32;
@@ -24,12 +24,12 @@ void param_init_u16(char* n_var0, uint16_t* n_var1);
 void param_init_u32(char* n_var0, uint32_t* n_var1);
 void param_init_float(char* n_var0, float* n_var1);
 struct param_info* param_get_by_name(const char* n_var0);
-struct param_info* param_get_by_index(uint16_t n_var0);
+struct param_info* param_get_by_index(int32_t n_var0);
 struct param_info* param_get_requested();
 float param_get_float_value(struct param_info* n_var0);
 void param_set_float_value(struct param_info* n_var0, float n_var1);
 extern struct param_info g_param_info[512U];
-extern uint16_t g_param_count;
+extern int32_t g_param_count;
 
 #ifdef __cplusplus
 }
