+++ smaccmpilot/src/gcs_transmit_driver.c
@@ -220,14 +220,14 @@ void gcs_transmit_send_param_value(struct param_info* n_var0,
     
     *&n_ref1->param_value = n_r2;
     
-    uint16_t* n_ref3 = &g_param_count;
-    uint16_t n_deref4 = *n_ref3;
+    int32_t* n_ref3 = &g_param_count;
+    int32_t n_deref4 = *n_ref3;
     
-    *&n_ref1->param_count = n_deref4;
+    *&n_ref1->param_count = (uint16_t) n_deref4;
     
-    uint16_t n_deref5 = *&n_var0->param_index;
+    int32_t n_deref5 = *&n_var0->param_index;
     
-    *&n_ref1->param_index = n_deref5;
+    *&n_ref1->param_index = (uint16_t) n_deref5;
     strncpy((uint8_t*) n_ref1->param_id, (const char*) n_var0->param_name, 16U);
     *&n_ref1->param_type = 0U;
     smavlink_send_param_value(n_ref1, n_var1, n_var2);
